fix(deploy): harden discovery builder, issuer, and build invocation
Move buildOpenIdConfiguration into a side-effect-free module so the build
script no longer triggers WebCrypto keypair generation. Treat empty
PORT/VERCEL_URL as unset. Use `pnpm exec tsx` in buildCommand. Refresh the
README bundling note.
