+++ apps/api-mock-deploy/scripts/generate-wellknown.ts
@@ -1,4 +1,4 @@
-import { buildOpenIdConfiguration } from "@zitadel/api-mock/server";
+import { buildOpenIdConfiguration } from "@zitadel/api-mock/openid-configuration";
 import { mkdir, writeFile } from "node:fs/promises";
 import { dirname, resolve } from "node:path";
 import { fileURLToPath } from "node:url";
